zero warnings

+++ Project/rose/source/rose_text.c
@@ -27,12 +27,12 @@ ROSE_Text* ROSE_TextCreate(const char* string) {
 		char character_index = (string[i] - 32); // todo: map oor chars to '?'
 		float w = (1.0f / 95.0f);
 		float x = (w * (float)character_index);
-		vertices[index + 0] = (ROSE_Vertex){ { i + 0, 0 }, { x,     1 } };
-		vertices[index + 1] = (ROSE_Vertex){ { i + 1, 0 }, { x + w, 1 } };
-		vertices[index + 2] = (ROSE_Vertex){ { i + 1,-1 }, { x + w, 0 } };
-		vertices[index + 3] = (ROSE_Vertex){ { i + 0, 0 }, { x,     1 } };
-		vertices[index + 4] = (ROSE_Vertex){ { i + 1,-1 }, { x + w, 0 } };
-		vertices[index + 5] = (ROSE_Vertex){ { i + 0,-1 }, { x,     0 } };
+		vertices[index + 0] = (ROSE_Vertex){ { (float)(i + 0),  0.0f }, { (float)(x),     1.0f } };
+		vertices[index + 1] = (ROSE_Vertex){ { (float)(i + 1),  0.0f }, { (float)(x + w), 1.0f } };
+		vertices[index + 2] = (ROSE_Vertex){ { (float)(i + 1), -1.0f }, { (float)(x + w), 0.0f } };
+		vertices[index + 3] = (ROSE_Vertex){ { (float)(i + 0),  0.0f }, { (float)(x),     1.0f } };
+		vertices[index + 4] = (ROSE_Vertex){ { (float)(i + 1), -1.0f }, { (float)(x + w), 0.0f } };
+		vertices[index + 5] = (ROSE_Vertex){ { (float)(i + 0), -1.0f }, { (float)(x),     0.0f } };
 	}
 
 	text->width = 7;
